Refactor the iterator based object listing to a baseclass
